% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromplot.R
\name{chromplot}
\alias{chromplot}
\title{Chromosome plot}
\usage{
chromplot(
  dat,
  annotation_thresh = NULL,
  title = "",
  label_all = 0,
  variant_list = NULL,
  size = 1.2,
  shape = 19,
  alpha = 1,
  color = c("darkblue", "#E69F00", "#00AFBB", "#999999", "#FC4E07", "darkorange1"),
  label_size = 3.5,
  sign_thresh = 5.1e-09,
  sign_thresh_color = "red",
  legend_position = "right",
  .checked = FALSE,
  show_xaxis = TRUE,
  axis_text_size = 11,
  title_text_size = 12,
  axis_title_size = 13,
  legend_title_size = 12,
  legend_text_size = 12,
  variant_ids_color = "red",
  legend_labels = NULL,
  legend_name = "Data",
  xmin = 0,
  highlight_genes = NULL,
  highlight_genes_color = NULL,
  variants = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  chr = NULL,
  vline = NULL,
  variant_ids = NULL
)
}
\arguments{
\item{annotation_thresh}{Display annotation for variants with p-values below this threshold}

\item{title}{Plot title (optional)}

\item{size}{Optional parameter setting the size of the plot points (default: \code{size=1.2})}

\item{shape}{A number of vector of numers setting the shape of the plotted points}

\item{alpha}{A number or vector of numbers setting the transparancy of the plotted points}

\item{color}{Optional parameter setting the color of the plot points (default: \code{color="darkblue"})}

\item{label_size}{Optional parameter to set the size of the plot labels (default: \code{label_size=3.5})}

\item{sign_thresh}{Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: \code{sign_thresh=5.1e-9}). Multiple thresholds can be provided in a vector, e.g \code{sign_thresh=c(5.1e-9,1.0e-6)}). Set this parameter to NULL if you dont want this line to appear at all \code{sign_thresh=NULL}}

\item{sign_thresh_color}{set the color of the significance threshold line or lines}

\item{legend_position}{Top,bottom,left or right}

\item{show_xaxis}{}

\item{axis_text_size}{Text size of the x and y axes tick labels (default: 12)}

\item{title_text_size}{Text size of the plot title (default: 13)}

\item{axis_title_size}{Text size of the x and y title labels (default: 12)}

\item{legend_title_size}{Text size of the legend title}

\item{legend_text_size}{Text size of the legend text}

\item{variant_ids_color}{The color of the variants in variants_id (default color is red)}

\item{legend_labels}{Legend labels}

\item{legend_name}{Change the name of the legend (default: None)}

\item{xmin, xmax}{Parameters setting the chromosomal range to display on the x-axis}

\item{highlight_genes}{A vector of genes or genes to highlight the datapoints for on the plot}

\item{highlight_genes_color}{Colors for the hihglighted genes (default: green)}

\item{variants}{Optional parameter, a dataset (required columns are \code{CHROM,POS,P}) with a list of variants to display and label on the plot.
Variants will be labeled with Gene_Symbol if present in the dataframe, otherwise with ID.  If neither Gene_Symbol nor ID are in the dataset, the variant will be labelled with its position (CHROM and POS)}

\item{ymin, ymax}{Optional parameters, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})}

\item{chr}{The chromosome to plot (i.e. chr15), only required if the input dataframe contains results from more than one chromosome}

\item{vline}{Parameter (optional) to add a vertical line to the plot at a specific chromosomal position, e.g \code{vline=204000066}. Multiple values can be provided in a vector, e.g  \code{vline=c(204000066,100500188)}}

\item{variant_ids}{A vector of variant ids to highlight on the plot, e.g. variant_ids=c("rs1234, rs45898")}

\item{df}{Dataframe (required columns are \code{CHROM,POS,P}), in upper- or lowercase) to use for the plot. See examples on how to retrieve the dataset with GOR.}

\item{rect}{Rectangle to add to the plot}

\item{rect_range}{Sets the size of the rectangle, default is 500 kb}

\item{variants_color}{One color or a vector of colors in case multiple variant dataframes are provided}

\item{variants_size}{A number or a vector of numbers (default: 1.2)}

\item{variants_alpha}{A number or a vector of numbers (default: 1)}

\item{variants_shape}{A number or a vector of numbers (default: 19)}

\item{geneplot_label_size}{Size of the labels for the genes on the gene and exon plots}
}
\value{
a chromosome plot (ggplot object)
}
\description{
\code{chromplot()} displays the association results for an entire chromosome.
Required parameter is a dataset (dataframe) containing the association data (with columns \code{CHROM,POS,P} in upper or lowercase)

All other input parameters are optional. A dataset including results from more than one
chromosome can be used as input if provided with the \code{chr parameter}.

E.g. \code{chromplot(multi_chr_dataset,chr = "chr15")}
}
\examples{


}
