% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectplot.R
\name{create_snpset}
\alias{create_snpset}
\title{Create a dataframe that can be used as input for making effect plots}
\usage{
create_snpset(
  df1,
  df2,
  thresh = 1e-08,
  protein_coding_only = TRUE,
  region_size = 1e+06,
  verbose = F
)
}
\arguments{
\item{df1}{The dataframe to extract the top snps from (with p-value below thresh)}

\item{df2}{The dataframe in which to search for overlapping SNPs from dataframe1}

\item{thresh}{The p-value threshold used for extracting the top snps from dataset 1}

\item{protein_coding_only}{Set this variable to TRUE to only use protein_coding genes for the annotation}

\item{region_size}{The size of the interval which to extract the top snps from}
}
\value{
Dataframe containing the top hit
}
\description{
\code{create_snpset()}
All other input parameters are optional
}
\examples{
\dontrun{
create_snpset(CD_UKBB,CD_FINNGEN, thresh=1e-09)
}

}
