% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectplot.R
\name{create_snpset_code}
\alias{create_snpset_code}
\title{Show the code/functions used to create a snpset}
\usage{
create_snpset_code()
}
\value{
Dataframe containing the top hit
}
\description{
\code{create_snpset_code()}
}
\examples{
\dontrun{
create_snpset_code()
}

}
