% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_checks.R
\name{dat_column_check_and_set}
\alias{dat_column_check_and_set}
\title{dat_column_check_and_set}
\usage{
dat_column_check_and_set(dat)
}
\arguments{
\item{dat}{A data frame or a list of data frames}
}
\description{
This function is used to standardize the column names in the input dataframe
}
