% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSGENES}
\alias{ENSGENES}
\title{Ensembl genes build 38}
\format{
A data frame  with 60,564 rows and 5 variables:
\describe{
\item{chrom}{Chromosome on build version 38 (GRCh38/hg38)}
\item{gene_start}{genetic position of gene start on build version 38}
\item{gene_end}{genetic position of gene end on build version 38 }
\item{gene_symbol}{The name of the gene}
\item{biotype}{the biotype of the gene}
}
}
\usage{
ENSGENES
}
\description{
Ensembl genes build 38
}
\keyword{datasets}
