% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectplot.R
\name{flip_to_positive_allele_for_dat1}
\alias{flip_to_positive_allele_for_dat1}
\title{Flip to the positive allele for dataset 1}
\usage{
flip_to_positive_allele_for_dat1(df)
}
\arguments{
\item{df}{A dataframe that is in the snpset format (like returned by the get_snpset() function)}
}
\value{
The input dataframe after flipping to the positive effect allele in dataframe 1
}
\description{
\code{flip_to_positive_allele_for_dat1()}
}
\examples{
\dontrun{
CD_UKBB_index_snps <- get_lead_snps(CD_UKBB)
snpset <- get_snpset(CD_UKBB_index_snps, CD_FINNGEN)
flip_to_positive_allele_for_dat1(snpset$matched)
}

}
