% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{get_gene}
\alias{get_gene}
\title{Get the genetic position of a gene by gene name}
\usage{
get_gene(gene_name, chr = NULL)
}
\arguments{
\item{gene_name}{A string representing a gene name (e.g. "FTO")}

\item{chr}{A string, search for the genes on this chromsome only, (e.g chr="chr1")}
}
\value{
Dataframe of genes
}
\description{
\code{get_gene()} Get the gene coordinates for a gene
Required parameter is gene name
}
\examples{
\dontrun{
get_gene("FTO")
}

}
