% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectplot.R
\name{get_snpset_code}
\alias{get_snpset_code}
\title{Show the code/functions used to get a snpset}
\usage{
get_snpset_code()
}
\value{
Dataframe containing the top hit
}
\description{
\code{get_snpset_code()}
}
\examples{
get_snpset_code()

}
