% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{get_top_hit}
\alias{get_top_hit}
\title{Get the top hit from the dataframe}
\usage{
get_top_hit(df, chr = NULL)
}
\arguments{
\item{df}{Dataframe containing association results}

\item{chr}{Get the top hit in the data frame for this chromosome. If chromosome is not provided, the top hit from the entire dataset is returned.}
}
\value{
Dataframe containing the top hit
}
\description{
\code{get_top_hit()} Get the top hit from the dataframe
All other input parameters are optional
}
\examples{
\dontrun{
data(gwas_CD)
get_top_hit(gwas_CD, chr="chr1")
}

}
