% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{get_top_snp}
\alias{get_top_snp}
\title{Get the top hit from the dataframe}
\usage{
get_top_snp(df, chr = NULL)
}
\arguments{
\item{df}{Dataframe containing association results}

\item{chr}{String, get the top hit in the data frame for this chromosome. If chromosome is not provided, the top hit from the entire dataset is returned.}
}
\value{
Dataframe containing the top hit
}
\description{
\code{get_top_snp()} Get the top hit from the dataframe
All other input parameters are optional
}
\examples{
\dontrun{
get_top_snp(CD_UKBB, chr="chr1")
}

}
