% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{get_topr_old_colors}
\alias{get_topr_old_colors}
\title{Get the topr custom colors from v. 2.0.2}
\usage{
get_topr_old_colors()
}
\value{
Vector of colors used for plotting
}
\description{
\code{get_topr_old_colors()} Get topr custom colors from v. 2.0.2
}
\examples{
\dontrun{
get_topr_old_colors()
}

}
