% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locuszoomplot.R
\name{locuszoom}
\alias{locuszoom}
\title{Locuszoom plot}
\usage{
locuszoom(
  df,
  ids2label,
  xmin = 0,
  xmax = max(df$POS),
  label_size = 3.5,
  genes = NULL,
  vline = NULL,
  sign_thresh = NULL
)
}
\arguments{
\item{xmin}{Parameters setting the chromosomal range to display on the x-axis}

\item{xmax}{Parameters setting the chromosomal range to display on the x-axis}

\item{label_size}{Optional parameter to set the size of the plot labels (default: \code{label_size=3.5})}

\item{genes}{List of genes or exons to plot}

\item{vline}{Parameter (optional) to add a vertical line to the plot at a specific chromosomal position, e.g \code{vline=204000066}. Multiple values can be provided in a vector, e.g  \code{vline=c(204000066,100500188)}}

\item{sign_thresh}{Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: \code{sign_thresh=5.1e-9}). Multiple thresholds can be provided in a vector, e.g \code{sign_thresh=c(5.1e-9,1.0e-6)}). Set this parameter to NULL if you dont want this line to appear at all \code{sign_thresh=NULL}}
}
\value{
ggplot object
}
\description{
\code{locuszoom()} displays the association results for a smaler region within one chromosome
Required parameter one dataset (dataframe) containing the association data (with columns \code{CHROM,POS,P,R2} in upper or lowercase), xmin and xmax

All other input parameters are optional
}
\examples{
\dontrun{
locuszoomt(df,xmin=xmin,xmax=xmax)
}
}
