% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectplot.R
\name{match_alleles}
\alias{match_alleles}
\title{Match the variants in the snpset by their alleles}
\usage{
match_alleles(df)
}
\arguments{
\item{df}{A dataframe that is in the snpset format (like returned by the get_snpset() function)}
}
\value{
The input dataframe containing only those variants whith matched alleles in the snpset
}
\description{
\code{matched_alleles()}
}
\examples{
\dontrun{
match_alleles(df)
}

}
