% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot_eddi.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ Plot}
\usage{
qq_plot(df, scale = 1, n_variants = 0, fontfamily = "", breaks = 15)
}
\arguments{
\item{df}{dataframe with required data}

\item{scale}{plot elements scale, default: 1}

\item{n_variants}{number of total variants used in the study}

\item{fontfamily}{font}
}
\value{
ggplot
}
\description{
QQ Plot
}
