% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{qqplot}
\alias{qqplot}
\title{QQ Plot}
\usage{
qqplot(df, scale = 1, n_variants = 0, fontfamily = "", breaks = 15)
}
\arguments{
\item{df}{dataframe with required data}

\item{scale}{plot elements scale, default: 1}

\item{n_variants}{number of total variants used in the study}

\item{fontfamily}{font}
}
\value{
ggplot
}
\description{
QQ Plot
}
