% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{R2_CD_UKBB}
\alias{R2_CD_UKBB}
\title{Example dataset including the R2 column for the locuszoom plot function}
\format{
A data frame  with 26,824 rows and 10 variables:
\describe{
\item{CHROM}{Chromosome, written as for example chr1 or 1}
\item{POS}{genetic position of the variant}
\item{ID}{Variant identifier, e.g. rsid }
\item{P}{P-value from Plink run, additive model, reggresion model GLM_FIRTH}
\item{R2}{variant correlation (r^2)}
}
}
\source{
A subset of the CD_UKBB dataset
}
\usage{
R2_CD_UKBB
}
\description{
The dataset is a subset of CD_UKBB and only includes variants above and near the IL23R gene on chromosome 1
}
\keyword{datasets}
