% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UC_UKBB}
\alias{UC_UKBB}
\title{UKBB Ulcerative colitis (ICD 10 code K51)}
\format{
A data frame  with 57,383 rows and 8 variables
\describe{
\item{CHROM}{Chromosome, written as for example chr1 or 1}
\item{POS}{genetic position of the variant}
\item{REF}{the reference allele}
\item{ALT}{the alternative allele}
\item{ID}{Variant identifier, e.g. rsid }
\item{P}{P-value from Plink run, additive model, regression model GLM_FIRTH}
\item{OR}{Odds Ratio}
\item{AF}{Allele frequency}
}
}
\source{
Ulcerative Colitis UKBB ICD10 code K51, only including variants with P<1e-03
}
\usage{
UC_UKBB
}
\description{
Dataset retrieved from the UK biobank including of 5,452 UC cases (K51) and 481,862 controls. The dataset has been filtered on variants with P<1e-03.
}
\keyword{datasets}
