% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CD_UKBB}
\alias{CD_UKBB}
\title{UKBB Crohns disease (ICD 10 code K50)}
\format{
A data frame  with 26,824 rows and 8 variables:
\describe{
\item{CHROM}{Chromosome, written as for example chr1 or 1}
\item{POS}{genetic position of the variant}
\item{REF}{the reference allele}
\item{ALT}{the alternative allele}
\item{ID}{Variant identifier, e.g. rsid }
\item{P}{P-value from Plink run, additive model, regression model GLM_FIRTH}
\item{OR}{Odds Ratio}
\item{AF}{Allele frequency}
}
}
\source{
Crohn's UKBB ICD10 code K50, only including variants with P<1e-03
}
\usage{
CD_UKBB
}
\description{
Dataset retrieved from the UK biobank consisting of 2,799 crohn´s cases (K50) and 484,515 controls. The dataset has been filtered on variants with P <1e-03.
}
\keyword{datasets}
