% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec_funcs.R
\name{create_snpset}
\alias{create_snpset}
\title{Create a dataframe that can be used as input for making effect plots}
\usage{
create_snpset(
  df1,
  df2,
  thresh = 1e-08,
  protein_coding_only = TRUE,
  region_size = 1e+06,
  verbose = F
)
}
\arguments{
\item{df1}{The dataframe to extract the top snps from (with p-value below thresh)}

\item{df2}{The dataframe in which to search for overlapping SNPs from dataframe1}

\item{thresh}{Numeric, the p-value threshold used for extracting the top snps from dataset 1}

\item{protein_coding_only}{Logical, set this variable to TRUE to only use protein_coding genes for the annotation}

\item{region_size}{Integer, the size of the interval which to extract the top snps from}

\item{verbose}{Logical, (default: FALSE). Assign to TRUE to get information on which alleles are matched and which are not.}
}
\value{
Dataframe containing the top hit
}
\description{
\code{create_snpset()}

This method is deprecated and will be removed in future versions. use \code{\link{get_snpset}} instead.
}
\examples{
\dontrun{
create_snpset(CD_UKBB,CD_FINNGEN, thresh=1e-09)
}

}
