% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec_funcs.R
\name{create_snpset_code}
\alias{create_snpset_code}
\title{Show the code/functions used to create a snpset}
\usage{
create_snpset_code()
}
\value{
Dataframe containing the top hit
}
\description{
This method is deprecated and will be removed in future versions. use \code{\link{get_snpset_code}} instead.

\code{create_snpset_code()}
}
\examples{
\dontrun{
create_snpset_code()
}

}
