% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ENSEXONS}
\alias{ENSEXONS}
\title{Ensembl exons build HG38-104-5-2}
\format{
A data frame  with 40,122 rows and 7 variables:
\describe{
\item{chrom}{Chromosome on build version 38 (GRCh38/hg38)}
\item{gene_start}{genetic position of gene start on build version 38}
\item{gene_end}{genetic position of gene end on build version 38 }
\item{gene_symbol}{The name of the gene}
\item{exon_chromstart}{genetic positions of exon start}
\item{exon_chromend}{genetic position of exon end}
}
}
\usage{
ENSEXONS
}
\description{
https://www.ensembl.info/2021/05/05/ensembl-104-has-been-released/
}
\details{
exons on chrY and chrM were excluded from the exon dataset
}
\keyword{datasets}
