% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{get_best_snp_per_MB}
\alias{get_best_snp_per_MB}
\title{Get the top SNP per 10 MB}
\usage{
get_best_snp_per_MB(
  df,
  thresh = 1e-09,
  region_size = 1e+06,
  protein_coding_only = FALSE,
  chr = NULL,
  .checked = FALSE,
  verbose = F
)
}
\arguments{
\item{df}{Dataframe}

\item{thresh}{A number.P -value threshold, only consider variants with p-values below this threshold (1e-09 by default)}

\item{region_size}{An integer (default = 1000000) indicating the window size for variant labelling. Increase this number for sparser annotation and decrease for denser annotation.}

\item{protein_coding_only}{Logical, set this variable to TRUE to only use protein_coding genes for annotation}

\item{chr}{String, get the top variants from one chromosome only, e.g. chr="chr1"}

\item{.checked}{Logical, if the input data has already been checked, this can be set to TRUE so it wont be checked again (FALSE by default)}

\item{verbose}{Logical, set to TRUE to get printed information on number of SNPs extracted}
}
\value{
Dataframe of lead variants. Returns the best variant per MB (by default, change the region size with the region argument) with p-values below the input threshold (thresh=1e-09 by default)
}
\description{
\code{get_best_snp_per_MB()} Get the top SNP per 1 MB region
All other input parameters are optional
}
\examples{
\dontrun{
get_best_snp_per_MB(CD_UKBB)
}

}
