% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{get_best_snp_per_MB}
\alias{get_best_snp_per_MB}
\title{Get the top SNP per 10 MB}
\usage{
get_best_snp_per_MB(df, thresh = 1e-09, region = 1e+06)
}
\arguments{
\item{df}{Dataframe}

\item{thresh}{Threshold}

\item{region}{region}
}
\value{
numeric vector
}
\description{
\code{get_best_snp_per_MB()} Get the top SNP per 10 MB
All other input parameters are optional
}
\examples{
\dontrun{
data(gwas_CD)
get_best_snp_per_MB(gwas_CD, thresh = 1e-09, region = 10000000)
}
}
