% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{get_gene}
\alias{get_gene}
\title{Get the genetic position of a gene or genes by their gene name}
\usage{
get_gene(gene_name, chr = NULL)
}
\arguments{
\item{gene_name}{string A gene name (e.g. "FTO")}

\item{chr}{Search for the gene on this chromosome}
}
\value{
Dataframe of genes
}
\description{
\code{get_gene()} Get gene coordinates by its gene symbol/name
Required parameters is on gene name or a vector of gene names
}
\examples{
\dontrun{
get_gene("FTO")
}

}
