% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec_funcs.R
\name{get_gene}
\alias{get_gene}
\title{Get the genetic position of a gene by gene name}
\usage{
get_gene(gene_name, chr = NULL, build = 38)
}
\arguments{
\item{gene_name}{A string representing a gene name (e.g. "FTO")}

\item{chr}{A string, search for the genes on this chromosome only, (e.g chr="chr1")}

\item{build}{A string, genome build, choose between builds 37 (GRCh37) and 38 (GRCh38) (default is 38)}
}
\value{
Dataframe with the gene name and its genetic coordinates
}
\description{
\code{get_gene()} Get the gene coordinates for a gene
Required parameter is gene name

This method is deprecated and will be removed in future versions. use \code{\link{get_gene_coords}} instead.
}
\examples{
\dontrun{
get_gene("FTO")
}
}
