% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setters_and_getters.R
\name{get_topr_colors}
\alias{get_topr_colors}
\title{Get the top hit from the dataframe}
\usage{
get_topr_colors()
}
\value{
Vector of colors used for plotting
}
\description{
\code{get_topr_colors()} Get the top hit from the dataframe
All other input parameters are optional
}
\examples{
\dontrun{
get_topr_colors()
}

}
