% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan plot}
\usage{
manhattan(
  df,
  ntop = 3,
  title = "",
  annotate = NULL,
  color = get_topr_colors(),
  sign_thresh = 5e-09,
  sign_thresh_color = "red",
  sign_thresh_label_size = 3.5,
  label_size = 3,
  size = 1,
  shape = 19,
  alpha = 1,
  highlight_genes_color = "green",
  highlight_genes_ypos = 1,
  axis_text_size = 11,
  axis_title_size = 12,
  title_text_size = 13,
  legend_title_size = 12,
  legend_text_size = 12,
  protein_coding_only = TRUE,
  angle = 0,
  legend_labels = NULL,
  gene = NULL,
  gene_padding = 1e+05,
  chr = NULL,
  annotate_with = "Gene_Symbol",
  region_size = 1e+06,
  legend_name = NULL,
  legend_position = "bottom",
  rect = NULL,
  nudge_x = 0.1,
  nudge_y = 0.2,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  highlight_genes = NULL,
  label_color = NULL
)
}
\arguments{
\item{df}{Dataframe or a list of dataframes (required columns are \code{CHROM,POS,P}), in upper- or lowercase) of association results.}

\item{ntop}{Number of datasets (GWASes) to show on the top plot}

\item{title}{Plot title (optional}

\item{annotate}{Display annotation for variants with p-values below this threshold}

\item{color}{Optional parameter setting the color of the plot points (default: \code{color="darkblue"})}

\item{sign_thresh}{Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: \code{sign_thresh=5.1e-9}). Multiple thresholds can be provided in a vector, e.g \code{sign_thresh=c(5.1e-9,1.0e-6)}). Set this parameter to NULL if you dont want this line to appear at all \code{sign_thresh=NULL}}

\item{sign_thresh_color}{set the color of the significance threshold line or lines}

\item{sign_thresh_label_size}{Set the text size of the label for the signficance thresholdds (default text sizze is 3.5)}

\item{label_size}{Optional parameter to set the size of the plot labels (default: \code{label_size=3})}

\item{size}{Optional parameter setting the size of the plot points (default: \code{size=1.2})}

\item{shape}{A number of vector of numers setting the shape of the plotted points}

\item{alpha}{A number or vector of numbers setting the transparancy of the plotted points}

\item{highlight_genes_color}{Colors for the hihglighted genes (default: green)}

\item{highlight_genes_ypos}{Display the genes at this position on the y-axis (default value is 1)}

\item{axis_text_size}{Text size of the x and y axes tick labels (default: 12)}

\item{axis_title_size}{Text size of the x and y title labels (default: 12)}

\item{title_text_size}{Text size of the plot title (default: 13)}

\item{legend_title_size}{Text size of the legend title}

\item{legend_text_size}{Text size of the legend text}

\item{protein_coding_only}{Set this parameter to TRUE to only use protein coding genes for annotation}

\item{angle}{The angle of the text label}

\item{legend_labels}{Legend labels}

\item{chr}{The chromosome to plot (i.e. chr15), only required if the input dataframe contains results from more than one chromosome}

\item{annotate_with}{Annotate the variants with eiher Gene_Symbol or ID (default annotate_with="Gene_Symbol")}

\item{region_size}{the size of the region used when annotating the top variant in a region (default value is 10000000 or 10 MB)}

\item{legend_name}{Change the name of the legend (default: None)}

\item{legend_position}{Top,bottom,left or right}

\item{rect}{Rectangle to add to the plot}

\item{nudge_x}{To vertically adjust the starting position of each gene label (this is a ggrepel parameter)}

\item{nudge_y}{To horizontally adjust the starting position of each gene label (this is a ggrepel parameter)}

\item{xmin, xmax}{Parameters setting the chromosomal range to display on the x-axis}

\item{ymin, ymax}{Optional parameters, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})}

\item{highlight_genes}{A vector of genes or genes to highlight the datapoints for on the plot}

\item{label_color}{Otpional parameter to change the color}
}
\value{
ggplot object
}
\description{
\code{manhattan()} displays the association results for the entire genome
Required parameter is at least one dataset (dataframe) containing the association data (with columns \code{CHROM,POS,P} in upper or lowercase)

All other input parameters are optional
}
\examples{
\dontrun{
data(gwas_CD)
manhattan(gwas_CD)
}

}
