% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{Manhattan plot}
\usage{
manhattan(
  df,
  annotation_thresh = 1e-09,
  ntop = 3,
  title = "",
  color = c("darkblue", "#E69F00", "#00AFBB", "#999999", "#FC4E07", "darkorange1"),
  sign_thresh = 5e-09,
  label_size = 3,
  size = 1,
  shape = 19,
  alpha = 1,
  highlight_genes_color = "green",
  variants_size = 1.2,
  variants_alpha = 1,
  variants_shape = 19,
  axis_text_size = 11,
  axis_title_size = 12,
  title_text_size = 13,
  legend_title_size = 12,
  legend_text_size = 12,
  rect = NULL,
  variant_list = NULL,
  variants = NULL,
  plot_order = NULL,
  legend_labels = NULL,
  legend_name = NULL,
  ymin = NULL,
  ymax = NULL,
  variants_color = NULL,
  highlight_genes = NULL,
  highlight_genes_ypos = NULL,
  legend_position = NULL
)
}
\arguments{
\item{df}{Dataframe (required columns are \code{CHROM,POS,P}), in upper- or lowercase) to use for the plot. See examples on how to retrieve the dataset with GOR.}

\item{annotation_thresh}{Display annotation for variants with p-values below this threshold}

\item{ntop}{Number of datasets (GWASes) to show on the top plot}

\item{title}{Plot title (optional)}

\item{color}{Optional parameter setting the color of the plot points (default: \code{color="darkblue"})}

\item{sign_thresh}{Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: \code{sign_thresh=5.1e-9}). Multiple thresholds can be provided in a vector, e.g \code{sign_thresh=c(5.1e-9,1.0e-6)}). Set this parameter to NULL if you dont want this line to appear at all \code{sign_thresh=NULL}}

\item{label_size}{Optional parameter to set the size of the plot labels (default: \code{label_size=3.5})}

\item{size}{Optional parameter setting the size of the plot points (default: \code{size=1.2})}

\item{shape}{A number of vector of numers setting the shape of the plotted points}

\item{alpha}{A number or vector of numbers setting the transparancy of the plotted points}

\item{highlight_genes_color}{Colors for the hihglighted genes (default: green)}

\item{variants_size}{A number or a vector of numbers (default: 1.2)}

\item{variants_alpha}{A number or a vector of numbers (default: 1)}

\item{variants_shape}{A number or a vector of numbers (default: 19)}

\item{axis_text_size}{Text size of the x and y axes tick labels (default: 12)}

\item{axis_title_size}{Text size of the x and y title labels (default: 12)}

\item{title_text_size}{Text size of the plot title (default: 13)}

\item{legend_title_size}{Text size of the legend title}

\item{legend_text_size}{Text size of the legend text}

\item{rect}{Rectangle to add to the plot}

\item{variants}{Optional parameter, a dataset (required columns are \code{CHROM,POS,P}) with a list of variants to display and label on the plot.
Variants will be labeled with Gene_Symbol if present in the dataframe, otherwise with ID.  If neither Gene_Symbol nor ID are in the dataset, the variant will be labelled with its position (CHROM and POS)}

\item{legend_labels}{Legend labels}

\item{legend_name}{Change the name of the legend (default: None)}

\item{ymin}{Optional parameters, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})}

\item{ymax}{Optional parameters, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})}

\item{variants_color}{One color or a vector of colors in case multiple variant dataframes are provided}

\item{highlight_genes}{A vector of genes or genes to highlight the datapoints for on the plot}

\item{legend_position}{Top,bottom,left or right}
}
\value{
plots using egg (https://cran.r-project.org/web/packages/egg/vignettes/Ecosystem.html)
}
\description{
\code{manhattan()} displays the association results for the entire genome
Required parameter is at least one dataset (dataframe) containing the association data (with columns \code{CHROM,POS,P} in upper or lowercase)

All other input parameters are optional
}
\examples{
\dontrun{
data(gwas_CD)
manhattan(gwas_CD)
}
}
