% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec_funcs.R
\name{match_alleles}
\alias{match_alleles}
\title{Match the variants in the snpset by their alleles}
\usage{
match_alleles(df, verbose = F)
}
\arguments{
\item{df}{A dataframe that is in the snpset format (like returned by the get_snpset() function)}

\item{verbose}{A logical scalar (default: FALSE). Assign to TRUE to get information on which alleles are matched and which are not.}
}
\value{
The input dataframe containing only those variants whith matched alleles in the snpset
}
\description{
\code{match_alleles()}

This method is deprecated and will be removed in future versions. use \code{\link{match_by_alleles}} instead.
}
\examples{
\dontrun{
match_alleles(df)
}

}
