% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionplot.R
\name{regionplot}
\alias{regionplot}
\title{Region plot}
\usage{
regionplot(
  df,
  ntop = 3,
  annotate = NULL,
  xmin = 0,
  size = 2,
  shape = 19,
  alpha = 1,
  label_size = 4,
  annotate_with = "ID",
  color = get_topr_colors(),
  axis_text_size = 11,
  axis_title_size = 12,
  title_text_size = 13,
  show_genes = FALSE,
  show_overview = TRUE,
  show_exons = FALSE,
  max_genes = 200,
  sign_thresh = 5e-09,
  sign_thresh_color = "red",
  sign_thresh_label_size = 3.5,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  protein_coding_only = FALSE,
  region = 1e+06,
  gene_region = 1e+05,
  angle = 0,
  legend_title_size = 12,
  legend_text_size = 12,
  nudge_x = 0.01,
  nudge_y = 0.01,
  rsids = NULL,
  rsids_color = "red",
  legend_name = "Data:",
  legend_position = "right",
  chr = NULL,
  vline = NULL,
  show_gene_names = NULL,
  legend_labels = NULL,
  gene = NULL,
  title = NULL
)
}
\arguments{
\item{df}{Dataframe or a list of dataframes (required columns are \code{CHROM,POS,P}), in upper- or lowercase) of association results.}

\item{ntop}{Number of datasets (GWASes) to show on the top plot}

\item{annotate}{Display annotation for variants with p-values below this threshold}

\item{xmin}{Parameters setting the chromosomal range to display on the x-axis}

\item{size}{Optional parameter setting the size of the plot points (default: \code{size=1.2})}

\item{shape}{A number of vector of numers setting the shape of the plotted points}

\item{alpha}{A number or vector of numbers setting the transparancy of the plotted points}

\item{label_size}{Optional parameter to set the size of the plot labels (default: \code{label_size=3.5})}

\item{annotate_with}{Annotate the variants with eiher Gene_Symbol or ID (default annotate_with="Gene_Symbol")}

\item{color}{Optional parameter setting the color of the plot points (default: \code{color="darkblue"})}

\item{axis_text_size}{Text size of the x and y axes tick labels (default: 12)}

\item{axis_title_size}{Text size of the x and y title labels (default: 12)}

\item{title_text_size}{Text size of the plot title (default: 13)}

\item{show_genes}{Show genes instead of exons (default show_genes=FALSE)}

\item{show_overview}{Show the overview plot (default show_overview=TRUE)}

\item{show_exons}{Show exons instead of genees (default show_exons=FALSE)}

\item{max_genes}{Only label the genes if they are fewer than max_genes (default values is 200).}

\item{sign_thresh}{Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: \code{sign_thresh=5.1e-9}). Multiple thresholds can be provided in a vector, e.g \code{sign_thresh=c(5.1e-9,1.0e-6)}). Set this parameter to NULL if you dont want this line to appear at all \code{sign_thresh=NULL}}

\item{sign_thresh_color}{set the color of the significance threshold line or lines}

\item{sign_thresh_label_size}{Set the text size of the label for the signficance thresholdds (default text sizze is 3.5)}

\item{xmax}{Parameters setting the chromosomal range to display on the x-axis}

\item{ymin}{Optional parameters, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})}

\item{ymax}{Optional parameters, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})}

\item{protein_coding_only}{Set this parameter to TRUE to only use protein coding genes for annotation}

\item{region}{the size of the region used when annotating the top variant in a region (default value is 10000000 or 10 MB)}

\item{gene_region}{The size of the region around the gene, if the gene argument was used (default: gene_region=100000)}

\item{angle}{The angle of the text label}

\item{legend_title_size}{Text size of the legend title}

\item{legend_text_size}{Text size of the legend text}

\item{nudge_x}{To vertically adjust the starting position of each gene label (this is a ggrepel parameter)}

\item{nudge_y}{To horizontally adjust the starting position of each gene label (this is a ggrepel parameter)}

\item{rsids}{A vector of rs ids to highlight on the plot, e.g. rsids=c("rs1234, rs45898")}

\item{rsids_color}{The color of the variants in variants_id (default color is red)}

\item{legend_name}{Change the name of the legend (default: None)}

\item{legend_position}{Top,bottom,left or right}

\item{chr}{The chromosome to plot (i.e. chr15), only required if the input dataframe contains results from more than one chromosome}

\item{vline}{Parameter (optional) to add a vertical line to the plot at a specific chromosomal position, e.g \code{vline=204000066}. Multiple values can be provided in a vector, e.g  \code{vline=c(204000066,100500188)}}

\item{show_gene_names}{Show the gene names even though they exceed the max_genes count}

\item{legend_labels}{Legend labels}

\item{gene}{Zoom in on this gene (e.g. gene=FTO)}

\item{title}{Plot title (optional}
}
\value{
plots using egg (https://cran.r-project.org/web/packages/egg/vignettes/Ecosystem.html)
}
\description{


\code{regionplot()} displays the association results for a smaler region within one chromosome
Required parameter is at least one dataset (dataframe) containing the association data (with columns \code{CHROM,POS,P} in upper or lowercase), xmin and xmax

All other input parameters are optional
}
\examples{
\dontrun{
regionplot(df,gene="FTO")
}
}
