% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionplot.R
\name{regionplot}
\alias{regionplot}
\title{Region plot}
\usage{
regionplot(
  dat,
  annotate = NULL,
  title = "",
  label_all = 0,
  xmin = 0,
  size = 2,
  shape = 19,
  alpha = 1,
  label_size = 4,
  color = c("darkblue", "#E69F00", "#00AFBB", "#999999", "#FC4E07", "darkorange1"),
  variants_color = "red",
  axis_text_size = 11,
  axis_title_size = 12,
  title_text_size = 13,
  legend_title_size = 12,
  legend_text_size = 12,
  show_genes = FALSE,
  show_overview = TRUE,
  show_exons = FALSE,
  sign_thresh = 5.1e-09,
  sign_thresh_color = "red",
  genes = NULL,
  variants = NULL,
  variant_ids = NULL,
  variant_ids_color = "red",
  xmax = NULL,
  ymin = NULL,
  ymax = NULL,
  protein_coding_only = FALSE,
  region = 1e+06,
  nudge_x = 0.01,
  nudge_y = 0.01,
  chr = NULL,
  vline = NULL,
  legend_name = "Data:",
  legend_position = "right",
  legend_labels = NULL,
  gene = NULL,
  highlight_genes = NULL,
  highlight_genes_color = NULL
)
}
\arguments{
\item{annotate}{Display annotation for variants with p-values below this threshold}

\item{title}{Plot title (optional)}

\item{xmin}{Parameters setting the chromosomal range to display on the x-axis}

\item{size}{Optional parameter setting the size of the plot points (default: \code{size=1.2})}

\item{shape}{A number of vector of numers setting the shape of the plotted points}

\item{alpha}{A number or vector of numbers setting the transparancy of the plotted points}

\item{label_size}{Optional parameter to set the size of the plot labels (default: \code{label_size=3.5})}

\item{color}{Optional parameter setting the color of the plot points (default: \code{color="darkblue"})}

\item{variants_color}{One color or a vector of colors in case multiple variant dataframes are provided}

\item{axis_text_size}{Text size of the x and y axes tick labels (default: 12)}

\item{axis_title_size}{Text size of the x and y title labels (default: 12)}

\item{title_text_size}{Text size of the plot title (default: 13)}

\item{legend_title_size}{Text size of the legend title}

\item{legend_text_size}{Text size of the legend text}

\item{show_genes}{Show genes instead of exons (default show_genes=FALSE)}

\item{show_overview}{Show the overview plot (default show_overview=TRUE)}

\item{show_exons}{Show exons instead of genees (default show_exons=FALSE)#'}

\item{sign_thresh}{Optional parameter setting the threshold of the dashed red horizontal line representing the significance threshold (default: \code{sign_thresh=5.1e-9}). Multiple thresholds can be provided in a vector, e.g \code{sign_thresh=c(5.1e-9,1.0e-6)}). Set this parameter to NULL if you dont want this line to appear at all \code{sign_thresh=NULL}}

\item{sign_thresh_color}{set the color of the significance threshold line or lines}

\item{genes}{List of genes or exons to plot}

\item{variants}{Optional parameter, a dataset (required columns are \code{CHROM,POS,P}) with a list of variants to display and label on the plot.
Variants will be labeled with Gene_Symbol if present in the dataframe, otherwise with ID.  If neither Gene_Symbol nor ID are in the dataset, the variant will be labelled with its position (CHROM and POS)}

\item{variant_ids}{A vector of variant ids to highlight on the plot, e.g. variant_ids=c("rs1234, rs45898")}

\item{variant_ids_color}{The color of the variants in variants_id (default color is red)}

\item{xmax}{Parameters setting the chromosomal range to display on the x-axis}

\item{ymin}{Optional parameters, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})}

\item{ymax}{Optional parameters, min and max of the y-axis, (default values: \code{ymin=0, ymax=max(-log10(df$P))})}

\item{protein_coding_only}{Set this parameter to TRUE to only use protein coding genes for annotation}

\item{region}{the size of the region used when annotating the top variant in a region (default value is 10000000 or 10 MB)}

\item{nudge_x}{To vertically adjust the starting position of each gene label (this is a ggrepel parameter)}

\item{nudge_y}{To horizontally adjust the starting position of each gene label (this is a ggrepel parameter)}

\item{chr}{The chromosome to plot (i.e. chr15), only required if the input dataframe contains results from more than one chromosome}

\item{vline}{Parameter (optional) to add a vertical line to the plot at a specific chromosomal position, e.g \code{vline=204000066}. Multiple values can be provided in a vector, e.g  \code{vline=c(204000066,100500188)}}

\item{legend_name}{Change the name of the legend (default: None)}

\item{legend_position}{Top,bottom,left or right}

\item{legend_labels}{Legend labels}

\item{highlight_genes}{A vector of genes or genes to highlight the datapoints for on the plot}

\item{highlight_genes_color}{Colors for the hihglighted genes (default: green)}
}
\value{
ggplot object
}
\description{


\code{regionplot()} displays the association results for a smaler region within one chromosome
Required parameter is at least one dataset (dataframe) containing the association data (with columns \code{CHROM,POS,P} in upper or lowercase), xmin and xmax

All other input parameters are optional
}
\examples{
\dontrun{
regionplot(df,xmin=xmin,xmax=xmax)
}
}
