% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topr.R
\docType{package}
\name{topr}
\alias{topr-package}
\alias{topr}
\title{topr}
\description{
A package for viewing and annotating genetic association data
}
\section{topr functions}{

The main plotting functions are:
\itemize{
\item \code{\link{manhattan}} to create Manhattan plot of association results
\item \code{\link{regionplot}} to create regional plots of association results for smaller genetic regions
}
}

\examples{
library(topr)
# Create a manhattan plot using
manhattan(CD_UKBB)

# Create a regional plot
regionplot(CD_UKBB, gene="IL23R")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/totajuliusd/topr}
  \item Report bugs at \url{https://github.com/totajuliusd/topr/issues}
}

}
\author{
\strong{Maintainer}: Thorhildur Juliusdottir \email{totajuliusd@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Andri Stefansson \email{andrimarstefansson@gmail.com}
}

Other contributors:
\itemize{
  \item Kyle Scott \email{kms309@miami.edu} [contributor]
}

}
